requirements.txt
beautifulsoup4==4.6.0
requests==2.18.4
progressbar2==3.34.3
